requirements.txt
Django==5.1.3
pillow==11.0.0

django-allauth==65.2.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2

flake8==7.1.1
isort==5.13.2

stripe==11.2.0

redis==5.2.0
celery==5.4.0
django-redis==5.4.0

psycopg2-binary==2.9.10
